+++ Client/Frontend/Settings/Main/About/LicenseAndAcknowledgementsSetting.swift
@@ -1,0 +1,22 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Shared
+
+/// Opens the license page in a new tab
+class LicenseAndAcknowledgementsSetting: Setting {
+    override var title: NSAttributedString? {
+        return NSAttributedString(string: .AppSettingsLicenses,
+                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
+    }
+
+    override var url: URL? {
+        return URL(string: "\(InternalURL.baseUrl)/\(AboutLicenseHandler.path)")
+    }
+
+    override func onClick(_ navigationController: UINavigationController?) {
+        setUpAndPushSettingsContentViewController(navigationController, self.url)
+    }
+}
