+++ Client/Frontend/Settings/Main/AppSettingsOptions.swift
@@ -8,98 +8,6 @@ import Shared
 import Account
 import LocalAuthentication
 
-// Show the current version of Firefox
-class VersionSetting: Setting {
-    unowned let settings: SettingsTableViewController
-
-    override var accessibilityIdentifier: String? { return "FxVersion" }
-    weak var appSettingsDelegate: AppSettingsDelegate?
-
-    init(settings: SettingsTableViewController,
-         appSettingsDelegate: AppSettingsDelegate) {
-        self.settings = settings
-        self.appSettingsDelegate = appSettingsDelegate
-        super.init(title: nil)
-    }
-
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: "\(AppName.shortName) \(AppInfo.appVersion) (\(AppInfo.buildNumber))", attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override func onConfigureCell(_ cell: UITableViewCell, theme: Theme) {
-        super.onConfigureCell(cell, theme: theme)
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        appSettingsDelegate?.clickedVersion()
-    }
-
-    override func onLongPress(_ navigationController: UINavigationController?) {
-        copyAppVersionAndPresentAlert(by: navigationController)
-    }
-
-    func copyAppVersionAndPresentAlert(by navigationController: UINavigationController?) {
-        let alertTitle: String = .SettingsCopyAppVersionAlertTitle
-        let alert = AlertController(title: alertTitle, message: nil, preferredStyle: .alert)
-        getSelectedCell(by: navigationController)?.setSelected(false, animated: true)
-        UIPasteboard.general.string = self.title?.string
-        navigationController?.topViewController?.present(alert, animated: true) {
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                alert.dismiss(animated: true)
-            }
-        }
-    }
-
-    func getSelectedCell(by navigationController: UINavigationController?) -> UITableViewCell? {
-        let controller = navigationController?.topViewController
-        let tableView = (controller as? AppSettingsTableViewController)?.tableView
-        guard let indexPath = tableView?.indexPathForSelectedRow else { return nil }
-        return tableView?.cellForRow(at: indexPath)
-    }
-}
-
-// Opens the license page in a new tab
-class LicenseAndAcknowledgementsSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsLicenses, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "\(InternalURL.baseUrl)/\(AboutLicenseHandler.path)")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-// Opens the App Store review page of this app
-class AppStoreReviewSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .Settings.About.RateOnAppStore, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        RatingPromptManager.goToAppStoreReview()
-    }
-}
-
-// Opens about:rights page in the content view controller
-class YourRightsSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsYourRights,
-                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "https://www.mozilla.org/about/legal/terms/firefox/")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
 // Opens the on-boarding screen again
 class ShowIntroductionSetting: Setting {
     let profile: Profile
